fuck

+++ src/it/vowch/android/adapters/VowAdapter.java
@@ -141,7 +141,8 @@ public class VowAdapter extends ArrayAdapter<ParseObject> {
 		} else {grade = "F"; 
 		}
 		
-		gradeView.setText(grade + " (" + successRate*100 + "%)");
+		
+		gradeView.setText(grade + " (" + (int) successRate*100 + "%)");
 		
 		/*
 		TextView timeLeftView = (TextView) rowView.findViewById(R.id.time_left);
